+++ tableauserverclient/server/request_options.py
@@ -108,14 +108,17 @@ class ImageRequestOptions(_FilterOptionsBase):
     class Resolution:
         High = 'high'
 
-    def __init__(self, imageresolution=None):
+    def __init__(self, imageresolution=None, maxage=None):
         super(ImageRequestOptions, self).__init__()
         self.image_resolution = imageresolution
+        self.max_age = maxage
 
     def apply_query_params(self, url):
         params = []
         if self.image_resolution:
             params.append('resolution={0}'.format(self.image_resolution))
+        if self.max_age:
+            params.append('maxAge={0}'.format(self.max_age))
 
         self._append_view_filters(params)